+++ example/lib/main.dart
@@ -2,19 +2,16 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:mvvm_plus/mvvm_plus.dart';
-import 'package:registrar/registrar.dart';
+import 'package:bilocator/bilocator.dart';
 
 void main() => runApp(myApp());
 
-Widget myApp() => Registrar<ColorService>(
-        builder: () => ColorService(milliSeconds: 1500),
-        child: MaterialApp(
-            debugShowCheckedModeBanner: false,
-            home: Registrar<ColorService>(
-              builder: () => ColorService(milliSeconds: 2250),
-              location: Location.tree,
-              child: Page(),
-            )));
+Widget myApp() => Bilocator<ColorService>(
+    builder: () => ColorService(milliSeconds: 1500),
+    child: MaterialApp(
+        debugShowCheckedModeBanner: false,
+        home: Bilocator<ColorService>(
+            builder: () => ColorService(milliSeconds: 2250), location: Location.tree, child: Page())));
 
 class IncrementButton extends View<IncrementButtonViewModel> {
   IncrementButton({super.key}) : super(builder: () => IncrementButtonViewModel());
@@ -23,18 +20,17 @@ class IncrementButton extends View<IncrementButtonViewModel> {
   Widget build(BuildContext context) {
     return FloatingActionButton(
       onPressed: viewModel.incrementCounter,
-      child: Text(viewModel.label, style: const TextStyle(fontSize: 24)),
+      child: Text(viewModel.buttonText, style: const TextStyle(fontSize: 24)),
     );
   }
 }
 
 class IncrementButtonViewModel extends ViewModel {
-  bool isNumber = false;
-  String get label => isNumber ? '+1' : '+a';
+  late final isNumber = createProperty<bool>(false);
+  String get buttonText => isNumber.value ? '+1' : '+a';
   void incrementCounter() {
-    isNumber ? get<PageViewModel>().incrementNumberCounter() : get<PageViewModel>().incrementLetterCounter();
-    isNumber = !isNumber;
-    buildView();
+    isNumber.value ? get<PageViewModel>().incrementNumber() : get<PageViewModel>().incrementLetter();
+    isNumber.value = !isNumber.value;
   }
 }
 
@@ -48,9 +44,9 @@ class Page extends View<PageViewModel> {
             child: Column(mainAxisAlignment: MainAxisAlignment.center, children: <Widget>[
           Row(mainAxisAlignment: MainAxisAlignment.center, children: <Widget>[
             Text(viewModel.letterCount.value,
-                style: TextStyle(fontSize: 64, color: listenTo<ColorService>(context: context).color)),
-            Text(viewModel.numberCounter.toString(),
-                style: TextStyle(fontSize: 64, color: listenTo<ColorService>().color)),
+                style: TextStyle(fontSize: 64, color: listenTo<ColorService>(context: context).color.value)),
+            Text(viewModel.numberCounter.value.toString(),
+                style: TextStyle(fontSize: 64, color: listenTo<ColorService>().color.value)),
           ]),
         ])),
         floatingActionButton: IncrementButton());
@@ -58,28 +54,23 @@ class Page extends View<PageViewModel> {
 }
 
 class PageViewModel extends ViewModel {
-  int numberCounter = 0;
-  late final letterCount = ValueNotifier<String>('a')..addListener(buildView);
+  late final numberCounter = createProperty<int>(0);
+  late final letterCount = createProperty<String>('a');
 
-  void incrementNumberCounter() {
-    numberCounter = numberCounter == 9 ? 0 : numberCounter + 1;
-    buildView();
-  }
-
-  void incrementLetterCounter() =>
+  void incrementNumber() => numberCounter.value = numberCounter.value == 9 ? 0 : numberCounter.value + 1;
+  void incrementLetter() =>
       letterCount.value = letterCount.value == 'z' ? 'a' : String.fromCharCode(letterCount.value.codeUnits[0] + 1);
 }
 
 class ColorService extends Model {
   ColorService({required int milliSeconds}) {
     _timer = Timer.periodic(Duration(milliseconds: milliSeconds), (_) {
-      color = <Color>[Colors.red, Colors.black, Colors.blue, Colors.orange][++_counter % 4];
-      notifyListeners();
+      color.value = <Color>[Colors.red, Colors.black, Colors.blue, Colors.orange][++_counter % 4];
     });
   }
 
   int _counter = 0;
-  Color color = Colors.orange;
+  late final color = createProperty<Color>(Colors.orange);
   late Timer _timer;
 
   @override